+++ packages/web/domain/common/types.ts
